requirements.txt
fastparquet==2024.11.0
kaggle==1.7.4.5
pandas==2.3.2
pymysql==1.1.2
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43